requirements.txt
fastapi
uvicorn
python-jsonstore
httpx==0.17.*
passlib[bcrypt]
basicauth
python-multipart
python-jose[cryptography]